// smart-lk/fixtures/lk-fixtures-type.ts
import { GraphQLClient } from 'graphql-request';
import { BaseLkPage } from '../pages/base-lk-page';
import { User } from '../../common/models/user';
import { UserWithCompany } from '../../common/models/user-with-company';

export type LkFixtures = {
  /**
   * - Возвращает объект класса BaseLkPage, включающий общие кастомные методы по работе с элементами на странице,
   *   а также методы, актуальные для Smart-Lk приложения
   */
  baseLkPage: BaseLkPage;
  /**
   * - Генерирует данные тестового пользователя
   * - Проводит регистрацию тестового пользователя по email (без доп. ролей)
   * - Возвращает тестового пользователя
   * - [После выполнения теста] Удаляет тестового пользователя
   */
  regularUser: User;
  /**
   * - Валидирует переданное через переменные окружения значение stage/dev1-4
   * - Возвращает URL Api Smart LK приложения для заданного stage/dev1-4 окружения
   */
  graphQLSmartLkApiUrl: string;
  /**
   * - Получает токен авторизации тестового пользователя
   * - Возвращает graphQl-клиента с установленными url (URL Api SmartLK приложения) и токеном
   */
  lkUserGrapqQLClient: GraphQLClient;
  /**
   * - [Вызывает фикстуру regularUser]
   * - Проходит онбординг тестовым пользователем
   * - Возвращает данные тестового пользователя и организации
   */
  onboardedUser: UserWithCompany;
  /**
   * - [Вызывает фикстуру regularUser]
   * - Генерирует данные организации для прохождения онбординга ЛК
   * - Возвращает данные прошедшего регистрацию тестового пользователя и организации
   */
  organizationOnboarding: UserWithCompany;
  /**
   * - Генерирует данные тестового пользователя c RU-префиксом номера телефона (Для тестов регистрации)
   * - Проводит регистрацию тестового пользователя по email (без доп. ролей)
   * - Возвращает тестового пользователя
   * - [После выполнения теста] Удаляет тестового пользователя
   */
  regularUserRuPhone: User;
  /**
   * - Генерирует данные тестового пользователя c BY-префиксом номера телефона (Для тестов регистрации)
   * - Проводит регистрацию тестового пользователя по email (без доп. ролей)
   * - Возвращает тестового пользователя
   * - [После выполнения теста] Удаляет тестового пользователя
   */
  regularUserByPhone: User;
  /**
   * - Генерирует данные тестового пользователя c KZ-префиксом номера телефона (Для тестов регистрации)
   * - Проводит регистрацию тестового пользователя по email (без доп. ролей)
   * - Возвращает тестового пользователя
   * - [После выполнения теста] Удаляет тестового пользователя
   */
  regularUserKzPhone: User;
  /**
   * - Генерирует данные тестового пользователя c RU-префиксом номера телефона
   * - Проводит регистрацию тестового пользователя по номеру телефона (без доп. ролей), без заполнения данных организации,
   *  без заполнения email
   * - Возвращает тестового пользователя
   * - [После выполнения теста] Удаляет тестового пользователя
   */
  registeredByRuPhoneUser: User;
};